requirements.txt
pytest==8.3.3
allure-pytest==2.13.5
selenium==4.0.0
Appium-Python-Client==2.0.0
